Implemented suggested change
Moved baseUrl and tags constants to blog-tags.js
